Move trigger map related tests to trigger_map_test.go and trigger_map_item_test.go
